Refactor code (#8)
* Made random thread safe

Made random thread safe in case I will need to use more threads for
animation in the game.

* Remove suo files from repository. No need to track them

* Minor fixes following code analysis run

1. Added Assembly version to OtheloUI + updated the version in both
assemblies to 1.0.0.1

2. main function is now marked with [STAThread] (following code warning
CA2232)

3. OtheloFormsUI now implemented the IDisposable interface

* Set .NET 4.51 + fixing designer for GameSettingsForm

Othelo now looks OK with .Net 4.51 + I refactored the GameSettingsForm
class. Now it is editable in the designer.

* Added Animation.cs

This class will be used to animate the game (fading text / sliding
tokens)

* Changed switching players notifications

Instead of a pop-up box, the name of the player will be shown directly
on the board.

* Made GameSettingsForm editable in WinForms Designer

* Made improvments to BoardForm

Made it editable in WinForms + Added animation to player names (instead
of the pop up )

* Animation implementation

Class used to animate text and tokens in the game

* Refactored files (using .net4 properties and other adjustments

* Refactored code

* Refactoring

1. Renamed TryPlacingATokenOnBoard to PlaceToken. This function should
always succeed.
2. Removed unnecessary allocations
3. Split the function GetBestPossibleMove to 2 functions:
a. GetPossibleMoves - return all moved available to the human player
b. GetBestComputerMove - return the computer move.
This caused a bit of duplicate code that I will sort in later versions.
For now, splitting this function was required to maintain SRP.

* Split Player to HumanPlayer and ComputerPlayer

Player is a parent of both. Separating the functionality between these
two classes made the code more maintainable and easier to understand.

* Part of the change to separating player to computer and human classes

* Create README.md (#7)